Updated storage tests to match changed behavior of save() method.
In waterline 0.11.0, the behavior of the save() method was changed to
no longer return the saved object.

Changed the tests to add a then() clause to pass along the pre-saved
object so it can be used to perform a delete(). This is pretty kludgy,
but it's more or less how the waterline folks modified their tests for
0.11.0, so it's in good company.